SIMA_AI_ENTERPRISE_V2/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
python-multipart==0.0.9
prometheus-client==0.20.0
psycopg[binary]==3.2.1
pypdf==4.3.1
httpx==0.27.2
paho-mqtt==2.1.0
qrcode==7.4.2
reportlab==4.2.2
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
